fix: Исправлен перенос товаров после покупки из корзины в совершенные заказы

+++ online-store/src/main/java/ru/goncharenko/market/order/dto/ItemInOrderDTO.java
@@ -14,6 +14,6 @@ import lombok.experimental.Accessors;
 public class ItemInOrderDTO {
 	long id;
 	String title;
-	long price;
+	Double price;
 	int count;
 }
