+++ online-store/src/main/java/ru/goncharenko/market/order/dto/OrderDTO.java
@@ -16,5 +16,5 @@ import java.util.List;
 public class OrderDTO {
 	private Long id;
 	private List<ItemInOrderDTO> items;
-	private Long totalSum;
+	private Double totalSum;
 }
